[#18,#19,#20] move status objects to domain
