+++ backend/api/src/main/kotlin/software/shonk/ConfigureApi.kt
@@ -5,7 +5,6 @@ import io.ktor.server.routing.route
 import io.ktor.server.routing.routing
 import software.shonk.adapters.incoming.configureHelloWorldController
 import software.shonk.adapters.incoming.configureShorkInterpreterController
-import software.shonk.adapters.outgoing.configureShorkInterpreterControllerOutgoing
 import software.shonk.application.service.HelloWorldService
 import software.shonk.application.service.MockShorkInterpreterService
 
@@ -14,10 +13,5 @@ fun Application.configureApi() {
     val shorkInterpreterUseCase = MockShorkInterpreterService()
 
     routing { route("/api") { configureHelloWorldController(helloWorldUseCase) } }
-    routing {
-        route("/api/v0") {
-            configureShorkInterpreterController(shorkInterpreterUseCase)
-            configureShorkInterpreterControllerOutgoing(shorkInterpreterUseCase)
-        }
-    }
+    routing { route("/api/v0") { configureShorkInterpreterController(shorkInterpreterUseCase) } }
 }
