+++ backend/api/src/main/kotlin/software/shonk/domain/GameState.kt
@@ -1,0 +1,10 @@
+package software.shonk.domain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class GameState {
+    NOT_STARTED,
+    RUNNING,
+    FINISHED,
+}
