+++ backend/api/src/main/kotlin/software/shonk/domain/Result.kt
@@ -1,0 +1,5 @@
+package software.shonk.domain
+
+import kotlinx.serialization.Serializable
+
+@Serializable data class Result(val winner: Winner)
