+++ backend/api/src/main/kotlin/software/shonk/domain/Status.kt
@@ -1,0 +1,11 @@
+package software.shonk.domain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Status(
+    val playerASubmitted: Boolean,
+    val playerBSubmitted: Boolean,
+    val gameState: GameState,
+    val result: Result,
+)
