+++ backend/api/src/main/kotlin/software/shonk/domain/Winner.kt
@@ -1,0 +1,10 @@
+package software.shonk.domain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class Winner {
+    A,
+    B,
+    UNDECIDED,
+}
